+++ packages/core/src/components/inputs/AddressInput/useAddressUtils.ts
@@ -1,0 +1,11 @@
+export function useAddressUtils() {
+  const isValidStreet = (value: string) => {
+    if (!value) {
+      return true
+    }
+    const regex = /^(.*)\s([0-9]+)(.*)$/
+    return regex.test(value)
+  }
+
+  return { isValidStreet }
+}