+++ packages/core/src/components/inputs/BaseInputV2/BaseInputV2.vue
@@ -1,26 +1,36 @@
 <template>
   <BaseInputWrapper>
-    <Field
-      as="input"
+    <input
       :id="name"
       :name="name"
+      v-model="value"
+      :class="{
+        dirty: meta.dirty,
+        valid: meta.touched && meta.valid,
+        invalid: meta.touched && !meta.valid
+      }"
       :type="type"
       v-bind="$attrs"
       placeholder="hidden"
+      @blur="handleBlur"
     />
     <label :for="name">{{ label }}</label>
-    <ErrorMessage class="error" :name="name" />
+    <span v-if="errorMessage && meta.touched" class="error">{{ errorMessage }}</span>
+    <template #postfix-icon>
+      <slot name="postfix-icon" />
+    </template>
   </BaseInputWrapper>
 </template>
-<script setup lang="ts">
-import type { RuleExpression } from 'vee-validate'
-import { ErrorMessage, Field } from 'vee-validate'
+<script setup lang="ts" generic="T">
+import { RuleExpression, useField } from 'vee-validate'
 import BaseInputWrapper from '@core/components/inputs/BaseInputWrapper/BaseInputWrapper.vue'
 
-type InputValue = string | number | null
-
-withDefaults(
+const props = withDefaults(
   defineProps<{
+    /**
+     * Used to identify this field in a form (VeeValidate Form).
+     */
+    modelValue?: T
     /**
      * Used to identify this field in a form (VeeValidate Form).
      */
@@ -32,7 +42,7 @@ withDefaults(
     /**
      * Validation constraints of this field, see https://vee-validate.logaretm.com/v4/api/use-field/#usage-with-typescript.
      */
-    validationRules?: RuleExpression<InputValue>
+    validationRules?: RuleExpression<T>
     /**
      * The type of the HTML `input` element, e.g. "text", "password", or "number".
      */
@@ -41,6 +51,13 @@ withDefaults(
   {
     type: 'text',
     validationRules: ''
+  }
+)
+const { value, meta, handleBlur, errorMessage } = useField<T>(
+  () => props.name,
+  props.validationRules,
+  {
+    syncVModel: true
   }
 )
 </script>