+++ packages/core/src/components/inputs/index.ts
@@ -1,33 +1,49 @@
+import AddressInput from '@core/components/inputs/AddressInput/AddressInput.vue'
 import BaseInput from '@core/components/inputs/BaseInput/BaseInput.vue'
 import BaseInputV2 from '@core/components/inputs/BaseInputV2/BaseInputV2.vue'
+import BaseInputV3 from '@core/components/inputs/BaseInputV3/BaseInputV3.vue'
 import BaseInputWrapper from '@core/components/inputs/BaseInputWrapper/BaseInputWrapper.vue'
 import BaseSelect from '@core/components/inputs/BaseSelect/BaseSelect.vue'
+import CentInput from '@core/components/inputs/CentInput/CentInput.vue'
 import CheckboxInput from '@core/components/inputs/CheckboxInput/CheckboxInput.vue'
-import CurrencyInput from '@core/components/inputs/CurrencyInput/CurrencyInput.vue'
+import CheckboxInputV2 from '@core/components/inputs/CheckboxInputV2/CheckboxInputV2.vue'
+import CodeInput from '@core/components/inputs/CodeInput/CodeInput.vue'
 import DateTimeInput from '@core/components/inputs/DateTimeInput/DateTimeInput.vue'
+import EuroInput from '@core/components/inputs/EuroInput/EuroInput.vue'
+import IbanInput from '@core/components/inputs/IbanInput/IbanInput.vue'
 import ImageInputWrapper from '@core/components/inputs/ImpageInputWrapper/ImageInputWrapper.vue'
 import LocationCoordinateSelect from '@core/components/inputs/LocationCoordinateSelect/LocationCoordinateSelect.vue'
+import MultiselectInput from '@core/components/inputs/MultiselectInput/MultiselectInput.vue'
 import NumberInput from '@core/components/inputs/NumberInput/NumberInput.vue'
 import PreviewImageInput from '@core/components/inputs/PreviewImageInput/PreviewImageInput.vue'
+import SearchInput from '@core/components/inputs/SearchInput/SearchInput.vue'
 import SelectInput from '@core/components/inputs/SelectInput/SelectInput.vue'
 import SliderToggleInput from '@core/components/inputs/SliderToggleInput/SliderToggleInput.vue'
+import SlugInput from '@core/components/inputs/SlugInput/SlugInput.vue'
 import UploadImageFAB from '@core/components/inputs/UploadImageFAB/UploadImageFAB.vue'
-import CodeInput from '@core/components/inputs/CodeInput/CodeInput.vue'
 
 export {
+  AddressInput,
   BaseInput,
   BaseInputV2,
+  BaseInputV3,
   BaseInputWrapper,
   BaseSelect,
+  CentInput,
   CheckboxInput,
-  CurrencyInput,
+  CheckboxInputV2,
   CodeInput,
   DateTimeInput,
+  EuroInput,
+  IbanInput,
   ImageInputWrapper,
   LocationCoordinateSelect,
+  MultiselectInput,
   NumberInput,
   PreviewImageInput,
+  SearchInput,
   SelectInput,
   SliderToggleInput,
+  SlugInput,
   UploadImageFAB
 }