+++ packages/gridstack/src/components/widgets/weather/composables/usePolygon.spec.ts
@@ -1,6 +1,4 @@
 import { describe, expect, it, vi } from 'vitest'
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-expect-error
 import booleanPointInPolygon from '@turf/boolean-point-in-polygon'
 import { usePolygon } from '@/components/widgets/weather/composables/usePolygon'
 import { GpsLocation } from '@lion5/component-library'